+++ TripDayFrontend/package.json
@@ -27,7 +27,7 @@
     "react-redux": "^9.1.2",
     "react-router-dom": "^6.23.1",
     "react-scripts": "^5.0.1",
-    "vite": "^5.2.12",
+    "vite": "^5.2.13",
     "web-vitals": "^4.0.1"
   },
   "scripts": {
