+++ TripDayFrontend/src/components/CustomMap.jsx
@@ -1,54 +1,19 @@
-import PropTypes from 'prop-types';
-import { useCallback } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import Map, { Marker, FullscreenControl, GeolocateControl, Source, Layer } from 'react-map-gl';
-import { setViewState, setMarker, setCurrentLocation, setSelectedPOI } from '../redux/reducers/mapReducer';
-import { MAPBOX_API_KEY } from '../constants/constants';
-
 import 'mapbox-gl/dist/mapbox-gl.css';
+import { useCallback, useRef } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import Map, { FullscreenControl, GeolocateControl } from 'react-map-gl';
+import { FoursquareResponsePropTypes } from '../constants/fourSqaurePropTypes';
+import { setViewState, setMarker, setCurrentLocation } from '../redux/reducers/mapReducer';
+import { MAPBOX_API_KEY } from '../constants/constants';
+// import ClickMarker from './ClickMarker';
+import ProximityMarkers from './ProximityMarkers';
+import AdditionalMarkerInfo from './AdditionalMarkerInfo';
 
-const createGeoJSONCircle = (center, radiusInMeters, points = 64) => {
-  const coords = {
-    longitude: center[0],
-    latitude: center[1],
-  };
-
-  const km = radiusInMeters / 1000;
-
-  const ret = [];
-  const distanceX = km / (111.320 * Math.cos((coords.latitude * Math.PI) / 180));
-  const distanceY = km / 110.574;
-
-  let theta;
-  let x;
-  let y;
-
-  for (let i = 0; i < points; i++) {
-    theta = (i / points) * (2 * Math.PI);
-    x = distanceX * Math.cos(theta);
-    y = distanceY * Math.sin(theta);
-
-    ret.push([coords.longitude + x, coords.latitude + y]);
-  }
-  ret.push(ret[0]);
-
-  return {
-    type: 'Feature',
-    geometry: {
-      type: 'Polygon',
-      coordinates: [ret],
-    },
-  };
-};
-
-const drawRadius = (lng, lat) => createGeoJSONCircle([lng, lat], 500);
-
+// react-map-gl component
 export default function CustomMap({ data }) {
   const mapStyle = useSelector((state) => state.mapReducer.mapStyle);
   const viewState = useSelector((state) => state.mapReducer.viewState);
-  const mapMarkers = useSelector((state) => state.mapReducer.markers);
-  const selectedPOI = useSelector((state) => state.mapReducer.selectedPOI);
-  const selectedPOIIcon = useSelector((state) => state.mapReducer.selectedPOIIcon);
+  const mapRef = useRef();
   const dispatch = useDispatch();
 
   const onMove = useCallback((event) => {
@@ -74,72 +39,9 @@ export default function CustomMap({ data }) {
     dispatch(setMarker(newMarker));
   };
 
-  const getClickMarker = () => ((mapMarkers.length > 0) ? mapMarkers.map((marker) => (
-    <div key={marker.id}>
-      <Marker longitude={marker.lng} latitude={marker.lat}>
-        <div className='text-4xl'>📍</div>
-      </Marker>
-      <Marker longitude={marker.lng} latitude={marker.lat} offset={[0, 30]}>
-        <div className='text-2xl  text-red-600'>{marker.restaurantName}</div>
-      </Marker>
-
-      <Source id='circle-data' type='geojson' data={drawRadius(marker.lng, marker.lat)}>
-        <Layer
-          id='circle-layer'
-          type='fill'
-          paint={{
-            'fill-color': '#777',
-            'fill-opacity': 0.3,
-          }}
-        />
-      </Source>
-    </div>
-  )) : null);
-
-  const getAdditionalMarkerInfo = () => {
-    if (data && data.results.length > 0) {
-      const filteredResult = data.results.filter((marker, i) => `${i + 1} ${marker.name}` === selectedPOI)[0];
-      if (filteredResult) {
-        return (
-          <div key={filteredResult.fsq_id}>
-            <Marker
-              longitude={filteredResult.geocodes.main.longitude}
-              latitude={filteredResult.geocodes.main.latitude}
-              offset={[0, -40]}
-            >
-              <div className='cardPOIAddInfo text-2xl  text-orange-400'>{`${filteredResult.name}`}</div>
-            </Marker>
-          </div>
-        );
-      }
-    }
-    return null;
-  };
-
-  const handlePOIMarkerClick = (event) => {
-    dispatch(setSelectedPOI(event.target._element.innerText));
-  };
-
-  const getProximityMarkers = () => (
-    ((data && data.results.length > 0) ? data.results.map((marker, i) => (
-      <div key={marker.fsq_id}>
-        <Marker longitude={marker.geocodes.main.longitude} latitude={marker.geocodes.main.latitude}>
-          <div className='text-4xl'>{selectedPOIIcon}</div>
-        </Marker>
-        <Marker
-          onClick={handlePOIMarkerClick}
-          longitude={marker.geocodes.main.longitude}
-          latitude={marker.geocodes.main.latitude}
-          offset={[0, 40]}
-        >
-          <div className='cardPOIMarker text-2xl  text-orange-400'>{`${i + 1} ${marker.name}`}</div>
-        </Marker>
-      </div>
-    )) : null)
-  );
-
   return (
     <Map
+      ref={mapRef}
       {...viewState}
       onMove={onMove}
       onClick={handleClick}
@@ -150,61 +52,12 @@ export default function CustomMap({ data }) {
     >
       <GeolocateControl positionOptions={{ enableHighAccuracy: true }} onGeolocate={handleCurrentLocation} showUserHeading />
       <FullscreenControl position='top-left' />
-      {getClickMarker()}
-      {getProximityMarkers()}
-      {getAdditionalMarkerInfo()}
+      {/* {mapRef ? <ClickMarker /> : null} */}
+      <ProximityMarkers data={data} />
+      <AdditionalMarkerInfo data={data} />
     </Map>
   );
 }
-
-const LocationPropType = PropTypes.shape({
-  address: PropTypes.string,
-  census_block: PropTypes.string,
-  country: PropTypes.string,
-  cross_street: PropTypes.string,
-  formatted_address: PropTypes.string,
-  locality: PropTypes.string,
-  postcode: PropTypes.string,
-  region: PropTypes.string,
-});
-
-const GeocodePropType = PropTypes.shape({
-  latitude: PropTypes.number,
-  longitude: PropTypes.number,
-});
-
-const CategoryPropType = PropTypes.shape({
-  id: PropTypes.number,
-  name: PropTypes.string,
-  icon: PropTypes.shape({
-    prefix: PropTypes.string,
-    suffix: PropTypes.string,
-  }),
-});
-
-const ChainPropType = PropTypes.shape({
-  id: PropTypes.string,
-  name: PropTypes.string,
-});
-
-const ResultPropType = PropTypes.shape({
-  fsq_id: PropTypes.string,
-  categories: PropTypes.arrayOf(CategoryPropType),
-  chains: PropTypes.arrayOf(ChainPropType),
-  distance: PropTypes.number,
-  geocodes: PropTypes.shape({
-    main: GeocodePropType,
-    roof: GeocodePropType,
-  }),
-  link: PropTypes.string,
-  location: LocationPropType,
-  name: PropTypes.string,
-  timezone: PropTypes.string,
-});
-
-const FoursquareResponsePropTypes = PropTypes.shape({
-  results: PropTypes.arrayOf(ResultPropType),
-});
 
 CustomMap.propTypes = {
   data: FoursquareResponsePropTypes,
